requirements.txt
numpy
sqlalchemy
pillow
fastapi[all]
openai
langchain
gradio
python-jose
passlib
python-multipart
email-validator
bcrypt
pre-commit